// src/src2/api/models/KernelAgencyResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type KernelAgencyResponse = {
    id?: string;
    organizationId?: string;
    code?: string;
    name?: string;
    location?: string;
    city?: string;
    country?: string;
    phone?: string;
    email?: string;
    agencyType?: string;
    isHeadquarter?: boolean;
    active?: boolean;
};